renamed files + fixed map component
