+++ package.json
@@ -9,6 +9,7 @@
     "@iconify/react": "^1.1.3",
     "@react-google-maps/api": "^1.13.0",
     "axios": "^0.21.0",
+    "classnames": "^2.2.6",
     "dotenv": "^8.2.0",
     "google-map-react": "^2.1.9",
     "nyc": "^15.1.0",
