--- src/component/Maps/maps.scss
@@ -1,39 +1,0 @@
-.Map {
-  width: 100%;
-  height: 100%;
-}
-
-.LocationPin {
-  //color: var(--main-blue);
-  position: fixed;
-  transform: translate(-50%, -86%);
-}
-
-.LocationPin__icon1 {
-  color: red;
-  font-size: 5rem;
-}
-
-.LocationPin__icon2 {
-  color: blue;
-  font-size: 5rem;
-}
-
-.LocationPin__text {
-  font-size: 1.3em;
-}
-
-@media only screen and (min-width: 768px) {
-  .Map {
-    margin: auto;
-    height: 100%;
-    width: 100%;
-  }
-}
-
-@media only screen and (max-width: 767px) {
-  .Map {
-    height: 40vh;
-    width: auto;
-  }
-}
